Use size_t instead of int in cset 073fbe7e6a3a.
* file-ops.cc (native_separator_path): Return type of length() on C++ string
is size_t, rather than int.

+++ liboctave/system/file-ops.cc
@@ -363,22 +363,24 @@ file_ops::concat (const std::string& dir, const std::string& file)
 }
 
 std::string
-file_ops::native_separator_path(const std::string& path)
+file_ops::native_separator_path (const std::string& path)
 {
   std::string retval;
-  if (dir_sep_char() == '/')
+
+  if (dir_sep_char () == '/')
     retval = path;
   else
     {
-      int n = path.length ();
-      for (int  i = 0; i < n; i++)
+      size_t n = path.length ();
+      for (size_t i = 0; i < n; i++)
         {
-          if (path[i] == '/') 
+          if (path[i] == '/')
             retval += dir_sep_char();
           else
             retval += path[i];
         }
     }
+
   return retval;
 }
